tasseo: add method for calculating max percent
Since cpu is added up for all hosts, we need to know the maximum
number available (i.e. the number of nodes, times 100% per host).

Add a helper function to get the number of measurements (find_num),
and use it to calculate the average number of nodes from cpu_user.